feat(profilePage): added settings page from profile page
