+++ docker-compose.yml
@@ -17,22 +17,22 @@ services:
       interval: 10s
       timeout: 5s
       retries: 5
-  aad-admin:
-    image: ghcr.io/stevensblueprint/aad-admin:latest
-    restart: on-failure
-    labels:
-      - "com.centurylinklabs.watchtower.scope=aad"
-    environment:
-      POSTGRES_USER: ${POSTGRES_USER}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
-      POSTGRES_DB: ${POSTGRES_DB}
-      DATABASE_URL: ${POSTGRES_URL}
-      NEXTAUTH_URL: ${NEXTAUTH_URL}
-      NEXTAUTH_SECRET: ${NEXTAUTH_SECRET}
-      GOOGLE_CLIENT_ID: ${GOOGLE_CLIENT_ID}
-      GOOGLE_CLIENT_SECRET: ${GOOGLE_CLIENT_SECRET}
-    ports:
-      - "3000:3000"
+  # aad-admin:
+  #   image: ghcr.io/stevensblueprint/aad-admin:latest
+  #   restart: on-failure
+  #   labels:
+  #     - "com.centurylinklabs.watchtower.scope=aad"
+  #   environment:
+  #     POSTGRES_USER: ${POSTGRES_USER}
+  #     POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
+  #     POSTGRES_DB: ${POSTGRES_DB}
+  #     DATABASE_URL: ${POSTGRES_URL}
+  #     NEXTAUTH_URL: ${NEXTAUTH_URL}
+  #     NEXTAUTH_SECRET: ${NEXTAUTH_SECRET}
+  #     GOOGLE_CLIENT_ID: ${GOOGLE_CLIENT_ID}
+  #     GOOGLE_CLIENT_SECRET: ${GOOGLE_CLIENT_SECRET}
+  #   ports:
+  #     - "3000:3000"
   wt:
     image: containrrr/watchtower
     labels:
