+++ src/pages/index.tsx
@@ -35,11 +35,6 @@ export default function Home() {
 function AuthShowcase() {
   const { data: sessionData } = useSession();
 
-  // const { data: secretMessage } = api.example.getSecretMessage.useQuery(
-  //   undefined, // no input
-  //   { enabled: sessionData?.user !== undefined }
-  // );
-
   return (
     <div className="mb-20 flex flex-col items-center justify-center gap-4">
       <p className="text-center text-2xl text-white">
@@ -52,9 +47,11 @@ function AuthShowcase() {
         {sessionData ? "Sign out" : "Sign in"}
       </Button>
       {sessionData && (
-        <Button variant="outlined" href={`/profiles/${sessionData?.user?.id}`}>
-          {"View Profile"}
-        </Button>
+        <>
+          <Button variant="outlined" href={`/profiles/${sessionData?.user?.id}`}>
+            {"View Profile"}
+          </Button>
+        </>
       )}
     </div>
   );
